Move numeric intrinsics into their own module
